+++ apps/fluent-tester/src/FluentTester/FluentTester.tsx
@@ -1,7 +1,6 @@
 import { Theme } from '@fluentui-react-native/framework';
-import { FocusTrapZone, Separator } from '@fluentui/react-native';
+import { FocusTrapZone, Separator, Text } from '@fluentui/react-native';
 import { Button } from '@fluentui-react-native/experimental-button';
-import { Text } from '@fluentui-react-native/experimental-text';
 import { themedStyleSheet } from '@fluentui-react-native/themed-stylesheet';
 import * as React from 'react';
 import { ScrollView, View, Text as RNText, Platform, SafeAreaView, BackHandler } from 'react-native';