+++ apps/fluent-tester/src/FluentTester/TestComponents/Menu/index.ts
@@ -1,0 +1,2 @@
+export * from './MenuTest';
+export * from './consts';