+++ packages/experimental/Menu/src/index.ts
@@ -1,0 +1,3 @@
+export { Menu } from './Menu/Menu';
+export { MenuTrigger } from './MenuTrigger/MenuTrigger';
+export { MenuPopover } from './MenuPopover/MenuPopover';