+++ packages/components/Callout/CHANGELOG.md
@@ -1,8 +1,34 @@
 # Change Log - @fluentui-react-native/callout
 
-This log was last generated on Wed, 13 Apr 2022 23:44:28 GMT and should not be manually modified.
+This log was last generated on Thu, 21 Apr 2022 21:50:03 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.19.43
+
+Thu, 21 Apr 2022 21:50:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.15.3
+- Bump @uifabricshared/foundation-compose to v1.11.17
+
+## 0.19.42
+
+Tue, 19 Apr 2022 16:54:55 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.15.2
+- Bump @uifabricshared/foundation-compose to v1.11.16
+
+## 0.19.41
+
+Mon, 18 Apr 2022 17:51:06 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.15.1
 
 ## 0.19.40