+++ packages/components/ContextualMenu/CHANGELOG.md
@@ -1,8 +1,47 @@
 # Change Log - @fluentui-react-native/contextual-menu
 
-This log was last generated on Wed, 13 Apr 2022 23:44:28 GMT and should not be manually modified.
+This log was last generated on Thu, 21 Apr 2022 21:50:03 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.16.24
+
+Thu, 21 Apr 2022 21:50:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.19.43
+- Bump @fluentui-react-native/focus-zone to v0.9.22
+- Bump @fluentui-react-native/icon to v0.11.18
+- Bump @fluentui-react-native/interactive-hooks to v0.15.3
+- Bump @fluentui-react-native/text to v0.11.19
+- Bump @uifabricshared/foundation-compose to v1.11.17
+- Bump @fluentui-react-native/pressable to v0.8.35
+
+## 0.16.23
+
+Tue, 19 Apr 2022 16:54:55 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.19.42
+- Bump @fluentui-react-native/focus-zone to v0.9.21
+- Bump @fluentui-react-native/icon to v0.11.17
+- Bump @fluentui-react-native/interactive-hooks to v0.15.2
+- Bump @fluentui-react-native/text to v0.11.18
+- Bump @uifabricshared/foundation-compose to v1.11.16
+- Bump @fluentui-react-native/pressable to v0.8.34
+
+## 0.16.22
+
+Mon, 18 Apr 2022 17:51:06 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.19.41
+- Bump @fluentui-react-native/focus-zone to v0.9.20
+- Bump @fluentui-react-native/interactive-hooks to v0.15.1
+- Bump @fluentui-react-native/pressable to v0.8.33
 
 ## 0.16.21