+++ packages/experimental/Avatar/CHANGELOG.md
@@ -1,8 +1,36 @@
 # Change Log - @fluentui-react-native/experimental-avatar
 
-This log was last generated on Wed, 06 Apr 2022 22:58:22 GMT and should not be manually modified.
+This log was last generated on Thu, 21 Apr 2022 21:50:03 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.14.15
+
+Thu, 21 Apr 2022 21:50:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/badge to v0.0.28
+- Bump @fluentui-react-native/framework to v0.7.23
+- Bump @fluentui-react-native/theme-tokens to v0.16.1
+
+## 0.14.14
+
+Wed, 20 Apr 2022 10:53:38 GMT
+
+### Patches
+
+- Replaced presence prop from PersonaCoin with the new PresenceBadge (v.kozova13@gmail.com)
+- Bump @fluentui-react-native/badge to v0.0.27
+
+## 0.14.13
+
+Tue, 19 Apr 2022 16:54:55 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.7.22
+- Bump @fluentui-react-native/theme-tokens to v0.16.0
 
 ## 0.14.12