+++ packages/experimental/Tabs/CHANGELOG.md
@@ -1,8 +1,41 @@
 # Change Log - @fluentui-react-native/experimental-tabs
 
-This log was last generated on Wed, 13 Apr 2022 23:44:28 GMT and should not be manually modified.
+This log was last generated on Thu, 21 Apr 2022 21:50:03 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.4.41
+
+Thu, 21 Apr 2022 21:50:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.7.23
+- Bump @fluentui-react-native/interactive-hooks to v0.15.3
+- Bump @fluentui-react-native/focus-zone to v0.9.22
+- Bump @fluentui-react-native/experimental-text to v0.7.23
+- Bump @fluentui-react-native/icon to v0.11.18
+
+## 0.4.40
+
+Tue, 19 Apr 2022 16:54:55 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.7.22
+- Bump @fluentui-react-native/interactive-hooks to v0.15.2
+- Bump @fluentui-react-native/focus-zone to v0.9.21
+- Bump @fluentui-react-native/experimental-text to v0.7.22
+- Bump @fluentui-react-native/icon to v0.11.17
+
+## 0.4.39
+
+Mon, 18 Apr 2022 17:51:06 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.15.1
+- Bump @fluentui-react-native/focus-zone to v0.9.20
 
 ## 0.4.38