+++ packages/theming/theme-tokens/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/theme-tokens
 
-This log was last generated on Wed, 06 Apr 2022 22:58:21 GMT and should not be manually modified.
+This log was last generated on Thu, 21 Apr 2022 21:50:03 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.16.1
+
+Thu, 21 Apr 2022 21:50:03 GMT
+
+### Patches
+
+- Bump design tokens packages (ruaraki@microsoft.com)
+
+## 0.16.0
+
+Tue, 19 Apr 2022 16:54:55 GMT
+
+### Minor changes
+
+- Move alias tokens to default theme (ruaraki@microsoft.com)
 
 ## 0.15.0